fix(columnar): widen segment IDs to u64 with overflow protection
Segment ID counters and delete-bitmap maps used u32, which would wrap
silently after ~4 billion segment allocations on a long-running node.
Widen everything to u64 and add explicit overflow protection.

- next_segment_id, memtable_segment_id: u32 → u64
- delete_bitmaps map key: u32 → u64
- on_memtable_flushed uses checked_add(1) and returns a typed error
  (SegmentIdExhausted) instead of wrapping silently
- All segment-ID parameters in public and crate-internal functions
  updated to u64
- Add regression test that triggers the error at u64::MAX

+++ nodedb-columnar/src/wal_record.rs
@@ -13,6 +13,7 @@ use zerompk::{FromMessagePack, ToMessagePack};
 
 /// A WAL record for a columnar collection operation.
 #[derive(Debug, Clone, Serialize, Deserialize, ToMessagePack, FromMessagePack)]
+#[non_exhaustive]
 pub enum ColumnarWalRecord {
     /// A row was inserted into the memtable.
     ///
@@ -33,7 +34,7 @@ pub enum ColumnarWalRecord {
     /// delete bitmap.
     DeleteRows {
         collection: String,
-        segment_id: u32,
+        segment_id: u64,
         row_indices: Vec<u32>,
     },
 
@@ -46,8 +47,8 @@ pub enum ColumnarWalRecord {
     ///   writing; discard and treat old segments as authoritative.
     CompactionCommit {
         collection: String,
-        old_segment_ids: Vec<u32>,
-        new_segment_ids: Vec<u32>,
+        old_segment_ids: Vec<u64>,
+        new_segment_ids: Vec<u64>,
     },
 
     /// The memtable was flushed to a new segment.
@@ -57,7 +58,7 @@ pub enum ColumnarWalRecord {
     /// the memtable via InsertRow records.
     MemtableFlushed {
         collection: String,
-        segment_id: u32,
+        segment_id: u64,
         row_count: u64,
     },
 }
